feat: Add support for custom logging function
This adds a `logger` property to the configuration options that takes a
method to use for logging rather than the default `console.log`.

People using a logging framework in their Express app can use this to
ensure the api-express-metrics logs are captured by their custom logger.

When no logger is specified, api-express-metrics will use `console.log`
as before.